Fix timing issues
There's a weird timing issue that can happen when the WebSocket connects
and fires the `connect` event before rAF fires. It's harmless, but
annoying. It causes the debugger to pause on the resulting exception.

This commit fixes it by ensuring that there is always something that
responds to `render` in the appropriate instance variable.